Struct gns3::GNS3Server

pub struct GNS3Server { /* fields omitted */ }

GNS3 Server Handle

Implementations

impl GNS3Server

pub fn new(address: impl AsRef<str>, port: u32) -> Result<Self, Error>

Create a new instance of a server handler

pub fn version(&self) -> &str

Get the version

pub fn get_projects(&self) -> Result<Vec<GNS3Project>, Error>

Returns all project informations

pub fn open_project(
&mut self,
project_id: impl AsRef<str>
) -> Result<GNS3Project, Error>

Opens a project, and returns the new information

pub fn close_project(&mut self) -> Result<GNS3Project, Error>

Opens a project, and returns the new information

pub fn create_project(
&mut self,
project_name: impl AsRef<str>
) -> Result<GNS3Project, Error>

Create a new project with the given name, and open it

pub fn get_templates(&self) -> Result<Vec<GNS3Template>, Error>

Returns all available templates

pub fn create_node(
&self,
name: impl AsRef<str>,
template_id: impl AsRef<str>
) -> Result<GNS3Node, Error>

Create a new node from a template

pub fn create_vpcs(&self, name: impl AsRef<str>) -> Result<GNS3Node, Error>

Create a new VPCS on the local compute

pub fn modify_node(
&self,
node_id: impl AsRef<str>,
name: Option<String>,
port: Option<u32>
) -> Result<GNS3Node, Error>

Modify the name or the port ID of a node

pub fn get_nodes(&self) -> Result<Vec<GNS3Node>, Error>

Return all nodes in the project

pub fn create_link(
&self,
node_a: &GNS3Node,
iface_a: usize,
node_b: &GNS3Node,
iface_b: usize
) -> Result<GNS3Link, Error>

Create a new link. The interface is used as an index into the interfaces of the corresponding node.

pub fn get_links(&self) -> Result<Vec<GNS3Link>, Error>

Return all links in the project

pub fn start_capture(&self, link: impl AsRef<str>) -> Result<GNS3Link, Error>

Start the capture on a specific link

pub fn stop_capture(&self, link: impl AsRef<str>) -> Result<GNS3Link, Error>

Stop the capture on a specific link

pub fn clear_capture_file(
&self,
link: impl AsRef<str>
) -> Result<GNS3Link, Error>

Stop the capture on a specific link

pub fn start_all_nodes(&self) -> Result<(), Error>

Start all nodes in the project

pub fn stop_all_nodes(&self) -> Result<(), Error>

Stop all nodes in the project

pub fn start_node(&self, node_id: impl AsRef<str>) -> Result<GNS3Node, Error>

Start a specific node

pub fn stop_node(&self, node_id: impl AsRef<str>) -> Result<GNS3Node, Error>

Stop a specific node

pub fn delete_project(
&mut self,
project_id: impl AsRef<str>
) -> Result<(), Error>

Delete an existing project